server/django/requirements.txt
acres==0.3.0
asgiref==3.8.1
boto3==1.38.18
botocore==1.38.18
certifi==2025.4.26
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.0
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
Django==4.2.1
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
etelemetry==0.3.1
filelock==3.18.0
httplib2==0.22.0
idna==3.10
isodate==0.6.1
jmespath==1.0.1
libretranslatepy==2.1.1
looseversion==1.3.0
lxml==5.4.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pathlib==1.0.1
pdf2image==1.17.0
pillow==11.2.1
prov==2.0.1
psycopg2-binary==2.9.10
puremagic==1.29
pydot==4.0.0
PyJWT==2.9.0
PyMuPDF==1.25.5
pyparsing==3.2.3
PyPDF2==3.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
pyxnat==1.6.3
rdflib==6.3.2
requests==2.32.3
s3transfer==0.12.0
scipy==1.15.3
simplejson==3.20.1
six==1.17.0
sqlparse==0.5.3
traits==7.0.2
translate==3.6.1
tzdata==2025.2
urllib3==2.4.0
